+++ src/feature_filter/convert.ts
@@ -28,7 +28,7 @@ type ExpectedTypes = {[_: string]: ExpressionInputType};
  *
  * We account for this by inserting a preflight type-checking expression before
  * each "any" term, allowing us to avoid evaluating the actual converted filter
- * if any type mismatches would cause it to produce an evalaution error:
+ * if any type mismatches would cause it to produce an evaluation error:
  *
  *     ["any",
  *       ["case",