+++ test/integration/expression/expression.test.ts
@@ -12,7 +12,6 @@ import {
 import ExpressionParsingError from '../../../src/expression/parsing_error';
 import {Result} from '../../../src/util/result';
 import {getGeometry} from '../../lib/geometry';
-import {stringify} from '../../lib/util';
 import {deepEqual, stripPrecision} from '../../lib/json-diff';
 
 const DECIMAL_SIGNIFICANT_FIGURES =  6;
@@ -52,7 +51,7 @@ describe('expression', () => {
                 delete fixture.metadata;
 
                 const fname = path.join(__dirname, expressionTestFileName);
-                fs.writeFileSync(fname, `${stringify(fixture)}\n`);
+                fs.writeFileSync(fname, JSON.stringify(fixture, null, 2));
                 return;
             }
 
@@ -62,13 +61,13 @@ describe('expression', () => {
             try {
                 expect(compileOk).toBeTruthy();
             } catch (e) {
-                throw new Error(`Compilation Falied:\nExpected ${JSON.stringify(expected.compiled)}\nResult   ${JSON.stringify(result.compiled)}`);
+                throw new Error(`Compilation Failed:\nExpected ${JSON.stringify(expected.compiled)}\nResult   ${JSON.stringify(result.compiled)}`);
             }
 
             try {
                 expect(evalOk).toBeTruthy();
             } catch (e) {
-                throw new Error(`Evaluation Falied:\nExpected ${JSON.stringify(expected.outputs)}\nResult   ${JSON.stringify(result.outputs)}`);
+                throw new Error(`Evaluation Failed:\nExpected ${JSON.stringify(expected.outputs)}\nResult   ${JSON.stringify(result.outputs)}`);
             }
 
         });